Merged PR 25301: Add PAYG account check in Transactional SMS admin
## What's being changed

This PR adds an account check to determine whether the EC account in scope is allowed to send SMS. If the check returns false, we disable the 'Enabled' Yes/No select and display a comment above. When doing this we also change the configured value for `transactional_sms/sms_settings/enabled` in the current scope, and in any direct child or parent scopes.

## Why it's being changed

SMS messages queued for accounts without that feature enabled in Engagement Cloud will still fail - this just informs the merchant and tidies up the configuration.

## How to review / test this change

- Erase ALL stored values for `transactional_sms/sms_settings/enabled`
- Ensure you have different websites and stores in your Magento, connected to some EC accounts WITH PAYG enabled, and others without the flag set
- Go to each different scope in the admin and test
- Is the message shown where required?
- Try enabling the config value for a scope you know doesn't support PAYG SMS. On load, is the value reset?
- If the value is forced to switch off at store view level, is the parent website also switched off?
- If the value is forced to switch off at website level, is the value also switched off for each child store?
- Can you hack it and break it by unchecking the default checkbox?

## Notes

- The frontend model handles the field's disabled state
- The backend model is required so that the value is correct after the switch off has been done (even if we reinit the config cache the change isn't reflected until the next page refresh).
- It is still possible to uncheck 'Use default' and remove the disabled state, however it's not possible to save 'Yes' if the account check fails.

Related work items: #122563