Merged PR 28113: Fix default sort order on SMS Sends Report
## What's being changed

Column sort order on the SMS Sends Report.

## Why it's being changed

We had almost all columns marked with `<sorting>asc<sorting>`, meaning that Magento had to pick one to start with. All columns are sortable by default, but the important bit is how the grid is sorted by default before the user interacts with it. Sorting by most recently created first seemed like a sensible default.

## How to review / test this change

- Go to the `ui_bookmark` table and remove any rows for `dotdigitalgroup_sms_report_grid`
- Clear cache and generated files
- Reload the report
- Confirm that rows are sorted by Created At, descending i.e. most recent first
- Confirm that other columns can still be sorted

## Notes

This will hopefully also fix occasional errors with the `TransactionalSmsOrderFlowTest` MFTF test, since that would fail if the top row in the table was not also the most recently created.

Related work items: #138624