Merged PR 34465: Protect against double-firing save_after events
## What's being changed

The `OrderSaveAfter` and `OrderCreditmemoSaveAfter` observers now use a service class to record that the observer has been executed, lest it be triggered again by a subsequent repeat event.

## Why it's being changed

We've not been able to replicate this locally, but one support ticket for the SMS module appears to indicate that in some setups, the `order_save_after` event is being fired twice - leading to duplicate SMS sends. We know from prior experience with other modules that observers can end up running twice or more; in the past we've used the Magento Registry to store a flag (although the service class approach is preferred).

We added the same code to `OrderCreditmemoSaveAfter`, not because a problem had been reported with that one, but for consistency and just in case it did occur.

## How to review / test this change

- In Magento\Framework\Model\AbstractModel duplicate line 829 (which dispatches all `xyz_save_after` events)
- Place an order
- Confirm that only one SMS row is created in the table
- Invoice the order
- Issue a credit memo for the order
- Again, confirm rows are created as expected

Related work items: #162701